pyproject.toml
[tool.poetry]
name = "aws-naip-tile-server"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"
packages = [
    {include='src'}
]

[tool.poetry.dependencies]
python = "^3.10"
rasterio = "^1.3.7"
Pillow = "^9.5.0"
shapely = "^2.0.1"
mercantile = "^1.2.1"
pyproj = "^3.5.0"
polars = "^0.18.0"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
isort = "^5.12.0"
pytest-dotenv = "^0.5.2"
requests = "^2.31.0"
flake8 = "^6.0.0"
flake8-black = "^0.3.6"
flake8-bugbear = "^23.6.5"
flake8-docstrings = "^1.7.0"
flake8-isort = "^6.0.0"
flake8-unused-arguments = "^0.0.13"
pytest = "^7.3.2"
coverage = {extras = ["toml"], version = "^7.2.7"}
boto3 = "^1.26.153"
awscli = "^1.27.153"
aws-sam-cli = "^1.86.1"
pre-commit = "^3.3.3"


[tool.isort]
profile = "black"

[tool.black]
line-length = 88
exclude = '''
^/(
  (
      \.aws-sam         # exclude a few common directories in the
  )/
  | foo.py           # also separately exclude a file named foo.py in
                     # the root of the project
)
'''

##### coverage configuration: https://coverage.readthedocs.io/en/latest/config.html
[tool.coverage]
[tool.coverage.paths]
source = ["src"]

[tool.coverage.run]
branch = true
source = ["src"]
omit = [".*", "tests/*"]

[tool.coverage.report]
show_missing = true
#fail_under = 100


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
